% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/coerce_call.R
\name{coerce_call}
\alias{coerce_call}
\title{Coerce call to hlme}
\usage{
coerce_call(call)
}
\arguments{
\item{call}{the call that is to be coerced}
}
\value{
will return another call whos arguments are combatitle with the hlme
function call
}
\description{
This will coerce a call so that the call arguments are valid arguments for
hlme function. If the formula arguments (fixed, random, mixture, classmb) are
character strings, they will be coerce to formula. If the argument data is a
character string, it will be evaluated as a symbol in the global env. The same goes
for the B argument
}
