% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cross_call.R
\name{cross_call}
\alias{cross_call}
\title{hlme argument combinations}
\usage{
cross_call(outcome, fixed_cov, age_var = NULL, ...)
}
\arguments{
\item{outcome}{character string or vector that specifies the left-hand side of the fixed
effects formula in the "fixed" hlme argument}

\item{age_var}{optional character string or vector that specifies the string
the character string to replace the substring "age_var" in the formula arguments
of hlme (fixed, random, mixture, classmb)}

\item{...}{other named arguments values or vectors to hlme function that must
represented as a string, numeric, or vector.}

\item{fix_cov}{character string or vector that specifies the right-hand sige
of the fixed effects formula in the "fixed" hlme argument}
}
\value{
will return a dataframe where the first three columns are the values
for outcome, fixed_cov, and age_var (if specified), and the remaining columns contain the
values hlme arguments whose columns correspond to the argument names.
}
\description{
Provide Arguments for hlme function and this function will create a dataframe
where each row contains a unique hlme call derived from the unique combinations
of argumnets
}
