% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/display_classes.R
\name{display_classes}
\alias{display_classes}
\title{Model Class Evaluation}
\usage{
display_classes(df, x, age, ...)
}
\arguments{
\item{df}{data frame used to produce model. create_pred_df output}

\item{x}{hlme model, or list of hlme models. If x is a list of hlme models,
they should be produced from the same data (df).}

\item{age}{string that specifies the age or time variable used in the model.
If polynomial terms were included, only the linear term should be specified.}

\item{...}{arguments to base plots (eg. main = "Title")}
}
\value{
kable output providing class analysis and plot output of classes
}
\description{
Provides n and frequency of each class and post prob of classes.
Also plots the model results using plot_hlme function
}
\examples{
library(dplyr)
library(ggplot2)

prep_data <- helphlme::prep_hlme2(lcmm::data_hlme,
                             vars = c("Time"),
                             center = FALSE,
                             scale = FALSE)

model1 <- helphlme::hlme2(
  data = prep_data,
  fixed = Y ~ Time + X1 + X2,
  mixture = ~ Time + X1 + X2,
  random = ~Time,
  subject = "ID",
  ng = 2
)

model2 <- helphlme::hlme2(
  data = prep_data,
  fixed = Y ~ Time + X1 + X2,
  mixture = ~ Time + X1 + X2,
  random = ~Time,
  subject = "ID",
  ng = 3
)

model_list <- list(model1 = model1, model2 = model2)

pred_df <- create_pred_df(prep_data,
                          age_vars = "Time",
                          fixcov = c("X1" = 0, "X2" = 0))

# works for single hlme models
display_classes(pred_df, model1, age = "Time")

# works for list of hlme models
display_classes(pred_df, model_list, age = "Time")
}
\author{
William F. Mueller
}
