% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{form_comb}
\alias{form_comb}
\title{Random Effects Formula Combinations}
\usage{
form_comb(cov)
}
\arguments{
\item{cov}{character vector of covariates}
}
\value{
character vector of random effects formula
}
\description{
Create Vector of all Random effect formula combination for lcmem given covariates
of interest
}
