% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/k_can.R
\name{k_can}
\alias{k_can}
\title{Class Number Candidates for Latent Class Mixed Effects Models}
\usage{
k_can(df, fixed, mixture, random, subject, max_k, idiag = FALSE, nwg = FALSE)
}
\arguments{
\item{df}{data frame object with data for models}

\item{fixed}{a string that represents a two-side linear formual object for the fixed effects in a linear mixed model.
By default, an intercept is included. If no intercept, -1 should be the first term included on the right of ~.}

\item{mixture}{a string that represents one-sided formula object for the class-specific fixed effects in the linear mixed model.
Among the list of covariates included in fixed,
the covariates with class-specific regression parameters are entered in mixture separated by +. By default, an intercept is included.
If no intercept, -1 should be the first term included.}

\item{random}{a string that represents an optional one-sided formula for the random-effects in the linear mixed model. Covariates
with a random-effect are separated by +. By default, an intercept is included. If no intercept, -1 should be the first term included.}

\item{subject}{name of the covariate representing the grouping structure specified with ''.}

\item{max_k}{the number of classes to apply the model structure}

\item{idiag}{optional logical for the structure of the variance-covariance matrix of the random-effects.
If FALSE, a non structured matrix of variance-covariance is considered (by default).
If TRUE a diagonal matrix of variance-covariance is considered.}

\item{nwg}{optional logical indicating if the variance-covariance of the random-effects is class-specific.
If FALSE the variance-covariance matrix is common over latent classes (by default).
If TRUE a class-specific proportional parameter multiplies the variance-covariance matrix in each class
(the proportional parameter in the last latent class equals 1 to ensure identifiability).}
}
\value{
a list that has lcmem output corresponding with the vector (1:max_k) provided.
}
\description{
This function applies lcmem across a vector of class candidates defined by (1:max_k)
and also allows the option to parallelize with the futures package
}
