% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gen_eval.R
\name{k_can_eval}
\alias{k_can_eval}
\title{Evaluate k_can models}
\usage{
k_can_eval(list)
}
\arguments{
\item{list}{List of lcmem outputs from k_can function}
}
\value{
data frame of with k, BIC and AIC
}
\description{
Outputs BIC and AIC for each model in the list provided
}
