% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/keval.R
\name{keval}
\alias{keval}
\title{Model Class Evaluation}
\usage{
keval(df, model, age, fixcov = NULL, main = "", xlab = "", ylab = "")
}
\arguments{
\item{df}{data frame used to produce model}

\item{model}{lcmem output to evaluate}

\item{age}{string of age variable}

\item{fixcov}{character vector specifiying other fixed effects covariates modelled besides age}

\item{main}{main title for plot}

\item{xlab}{x label for plot}

\item{ylab}{y label for plot}
}
\value{
kable output providing class analysis and plot output
}
\description{
Provides n and frequency of each class and post prob of classes.
Also plots the model results using lcplot function
}
