% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/keval.R
\name{ksum}
\alias{ksum}
\title{Model Class Evaluation Table}
\usage{
ksum(model)
}
\arguments{
\item{model}{hlme model object}
}
\value{
data frame with mean of post probability values, n, frequency.
}
\description{
Creates Dataframe evaluating each class of an hlme model
}
