% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/lcmem_prep.R
\name{lcmem_prep}
\alias{lcmem_prep}
\title{Prep Data Frame for trajpkg functions}
\usage{
lcmem_prep(df, vars, center = TRUE, scale = TRUE)
}
\arguments{
\item{df}{Dataframe object}

\item{vars}{character vector providing the names of the columns we would like to center and scale}

\item{center}{a boolean that specifies if vars should be centered
by subtracting by the mean}

\item{scale}{a boolean that spcifies if vars should be devided by their standard deviation}
}
\value{
a dataframe with vars centered and scaled and their nonscaled values stores as var_ns
}
\description{
This function centers and scales the variables provided by vars by subtracting by the mean
and dividing by the starndard deviation. It will also create new columns appending with
"_ns" to save the original nonscaled values.
}
