% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/lcplot.R
\name{lcplot}
\alias{lcplot}
\title{Plot the trajectories of a an hlme object}
\usage{
lcplot(df, model, age, ...)
}
\arguments{
\item{df}{data frame used to predict model outcomes. Should be prodcued by lcpred}

\item{model}{hlme model output. If modelled using lcmem, use lcmem$model.}

\item{age}{string specfiying age variable name}

\item{fixcov}{character vector specifiying other fixed effects covariates modelled besides age}
}
\value{
outputs base plot object. This object cannot be assigned to an R object.
}
\description{
Provide an hlme model, this function will use the base plot function to plot the trajectories
determined by the fixed effects of the model
}
