% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/lcplot.R
\name{lcplot_apply}
\alias{lcplot_apply}
\title{Plot the trajectories of a list of lcmem function outputs}
\usage{
lcplot_apply(
  df,
  model_list,
  age,
  fixcov = NULL,
  main = "",
  xlab = "",
  ylab = ""
)
}
\arguments{
\item{df}{data frame used to model the data. Should be created by lcmem_prep function}

\item{model_list}{List of lcmem outputs.}

\item{age}{string specfiying age variable name}

\item{fixcov}{character vector specifiying other fixed effects covariates modelled besides age}

\item{main}{main title for plot}

\item{xlab}{xlabel for plot}

\item{ylab}{for plot}
}
\value{
outputs base plot objects for each model provided.
}
\description{
Provided a list of lcmem function outputs, this function will use the base plot function
to plot the trajectories determined by the fixed effects of the model. Subtitle of the plots
will be automatically determined by k and random effects structure
}
