% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/lcplot.R
\name{lcplot_prep}
\alias{lcplot_prep}
\title{Lcplot Prep}
\usage{
lcplot_prep(
  df,
  age,
  fixcov = NULL,
  square = NULL,
  cube = NULL,
  center = FALSE,
  scale = FALSE
)
}
\arguments{
\item{df}{Dataframe from to create predictions produced by lcmem prep}

\item{age}{string that gives the name of the age covariate}

\item{fixcov}{character vector of fixed covariates that were given to the model}

\item{square}{name of quadradic age variable}

\item{cube}{name of cubic age variable}

\item{center}{did we center the age variables when modelling?}

\item{scale}{did we scale the age variables when modelling?}
}
\description{
THis function will create a dataframe to provide inputs to hlme model.
}
