% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/lcplot.R
\name{lcpred}
\alias{lcpred}
\title{Create Dataframe to Predict hlme Outcome}
\usage{
lcpred(
  df,
  age,
  square = NULL,
  cube = NULL,
  fixcov = NULL,
  center = FALSE,
  scale = FALSE
)
}
\arguments{
\item{df}{dataframe create used to model the data. Should be created by lcmem_prep fucntion}

\item{age}{string of age variable used in the model}

\item{square}{string of quadratic age variable if a quadratic term was include}

\item{cube}{character string of cubic age variable if cubic variable included}

\item{fixcov}{character vector of other factor covariates included in the model (eg. sex and strain)}

\item{center}{was data centered T or F}

\item{scale}{was data scale T or F}
}
\value{
a dataframe
}
\description{
This funciton creates dataframes that are used by the predictY function in the lcmm package to plot hlme models
}
