% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/lcmem_prep.R
\name{lcprep}
\alias{lcprep}
\title{Lcprep creates a class and preps dataframe}
\usage{
lcprep(df, age, square = NULL, cube = NULL, center = FALSE, scale = FALSE)
}
\arguments{
\item{df}{the dataframe to prep}

\item{age}{age variable to prep}

\item{square}{should aged be squared}

\item{cube}{should age be cubed?}

\item{center}{should age varaibles be centered}

\item{scale}{should age variable be scaled}
}
\value{
an object of lcprep class
}
\description{
prepares a dataframe for latent class trajectory modelling
}
