% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/lcresid_plot.R
\name{lcresid_plot}
\alias{lcresid_plot}
\title{Plot the residuals of hlme model by class}
\usage{
lcresid_plot(data, model, nameofoutcome, nameofage, ylimit = c(-5, 5))
}
\arguments{
\item{data}{data frame that was used to create the model}

\item{model}{hlme model output that we would like to plot}

\item{nameofoutcome}{string name of outcome variable}

\item{nameofage}{string name of age variable}

\item{ylimit}{numeric 2 element vector specifying y minimum and y maximum}
}
\value{
ggplot object of plotted residuals by class
}
\description{
This function will plot the residuals of an hlme model by class
}
