% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gen_eval.R
\name{mo_eval}
\alias{mo_eval}
\title{Evaluate models}
\usage{
mo_eval(list)
}
\arguments{
\item{list}{List of lcmem outputs from ran_can function}
}
\value{
data frame of with model number, model details, BIC and AIC other model selection criteria. This function
also returns plots of the model selection criteria by model number.
}
\description{
Outputs BIC and AIC for each model in the list provided
}
