% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_hlme.R
\name{plot_model}
\alias{plot_model}
\title{Plot the trajectories of a an hlme object}
\usage{
plot_model(df, model, age, ...)
}
\arguments{
\item{df}{data frame used to predict model outcomes. Should be prodcued by
create_pred_df function}

\item{model}{hlme model output. If modelled using lcmem, use lcmem$model.}

\item{age}{string specfiying age variable name}

\item{...}{arguements to pass to the plot function (eg. main = "title of plot")}
}
\value{
outputs base plot object. This object cannot be assigned to an R object.
}
\description{
Provide an hlme model, this function will use the base plot function to plot the trajectories
determined by the fixed effects of the model
}
