% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pmap_hlme2.R
\name{pmap_hlme2}
\alias{pmap_hlme2}
\title{Row (p) Map hlme2}
\usage{
pmap_hlme2(data)
}
\description{
This function will map a dataframe where each row provides a hlme2
function call whose arguments are specified by the column names. This
dataframe can be best created using the cross_call function
}
