% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ran_can.R
\name{ran_can}
\alias{ran_can}
\title{Random Effects Candidates for Latent Class Mixed Effects Models}
\usage{
ran_can(df, fixed, mixture, random_vect, subject, k = 5)
}
\arguments{
\item{df}{data frame object with data for models}

\item{fixed}{a string that represents a two-side linear formual object for the fixed effects in a linear mixed model.
By default, an intercept is included. If no intercept, -1 should be the first term included on the right of ~.}

\item{mixture}{a string that represents one-sided formula object for the class-specific fixed effects in the linear mixed model
(to specify only for a number of latent classes greater than 1). Among the list of covariates included in fixed,
the covariates with class-specific regression parameters are entered in mixture separated by +. By default, an intercept is included.
If no intercept, -1 should be the first term included.}

\item{random_vect}{a character vector where each element represents an optional one-sided formula for the random-effects in the linear mixed model. Covariates
with a random-effect are separated by +. By default, an intercept is included. If no intercept, -1 should be the first term included.}

\item{subject}{name of the covariate representing the grouping structure specified with ''.}

\item{k}{number of classes}
}
\value{
a list that has lcmem output corresponding with the random_vect provided.
}
\description{
This function applies lcmem across a vector of random effects candidates and also allows
the option to parallelize with the futures package
}
