% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gen_eval.R
\name{ran_can_eval}
\alias{ran_can_eval}
\title{Evaluate ran_can models}
\usage{
ran_can_eval(list)
}
\arguments{
\item{list}{List of lcmem outputs from ran_can function}
}
\value{
data frame of with model number, model details, BIC and AIC
}
\description{
Outputs BIC and AIC for each model in the list provided
}
