% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gen_eval.R
\name{ran_refine_eval}
\alias{ran_refine_eval}
\title{Evaluate ran_refine models}
\usage{
ran_refine_eval(list)
}
\arguments{
\item{list}{List of lcmem outputs from ran_refine function}
}
\value{
data frame of with model number, model details, BIC and AIC
}
\description{
Outputs BIC and AIC for each model in the list provided
}
