% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{sample_df}
\alias{sample_df}
\title{Sample Data Frame}
\usage{
sample_df(df, id, n)
}
\arguments{
\item{df}{data frame to sample}

\item{id}{identifier variable name as string}

\item{n}{number of ids to sample}
}
\value{
a data frame
}
\description{
Sample data frame, subsets n ids
}
