% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/display_classes.R
\name{table_classes}
\alias{table_classes}
\title{Model Class Evaluation Table}
\usage{
table_classes(model)
}
\arguments{
\item{model}{hlme model object}
}
\value{
data frame with mean of post probability values, n, frequency.
}
\description{
Creates Dataframe evaluating each class of an hlme model
}
