% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/display_classes.R
\name{display_classes}
\alias{display_classes}
\title{Model Class Evaluation}
\usage{
display_classes(df, x, age, ...)
}
\arguments{
\item{df}{data frame used to produce model. create_pred_df output}

\item{x}{hlme model, or list of hlme models. If x is a list of hlme models,
they should be produced from the same data (df).}

\item{age}{string that specifies the age or time variable used in the model.
If polynomial terms were included, only the linear term should be specified.}

\item{...}{arguments to base plots (eg. main = "Title")}
}
\value{
kable output providing class analysis and plot output of classes
}
\description{
Provides n and frequency of each class and post prob of classes.
Also plots the model results using plot_hlme function
}
\author{
William F. Mueller
}
