% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/k_can.R
\name{k_can}
\alias{k_can}
\title{Class Number Candidates for Latent Class Mixed Effects Models}
\usage{
k_can(df, fixed, mixture, random, subject, max_k)
}
\arguments{
\item{df}{data frame object with data for models}

\item{fixed}{a string that represents a two-side linear formual object for the fixed effects in a linear mixed model.
By default, an intercept is included. If no intercept, -1 should be the first term included on the right of ~.}

\item{mixture}{a string that represents one-sided formula object for the class-specific fixed effects in the linear mixed model
(to specify only for a number of latent classes greater than 1). Among the list of covariates included in fixed,
the covariates with class-specific regression parameters are entered in mixture separated by +. By default, an intercept is included.
If no intercept, -1 should be the first term included.}

\item{random}{a string that represents an optional one-sided formula for the random-effects in the linear mixed model. Covariates
with a random-effect are separated by +. By default, an intercept is included. If no intercept, -1 should be the first term included.}

\item{subject}{name of the covariate representing the grouping structure specified with ''.}

\item{max_k}{the number of classes to apply the model structure}
}
\value{
a list that has lcmem output corresponding with the vector (1:max_k) provided.
}
\description{
This function applies lcmem across a vector of class candidates defined by (1:max_k)
and also allows the option to parallelize with the futures package
}
