% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/keval.R
\name{keval}
\alias{keval}
\title{Model Class Evaluation}
\usage{
keval(df, model, age, ...)
}
\arguments{
\item{df}{data frame used to produce model. lcpred output}

\item{model}{lcmem output to evaluate}

\item{age}{string of age variable}

\item{...}{arguments to base plots (eg. main = "Title")}
}
\value{
kable output providing class analysis and plot output
}
\description{
Provides n and frequency of each class and post prob of classes.
Also plots the model results using lcplot function
}
\examples{
library(dplyr)
library(ggplot2)

prep_data <- helphlme::prep_hlme2(lcmm::data_hlme,
                             vars = c("Time"),
                             center = FALSE,
                             scale = FALSE)
model1 <- helphlme::hlme2(
  data = prep_data,
  fixed = Y ~ Time + X1 + X2,
  mixture = Y ~ Time + X1 + X2,
  random = ~Time,
  subject = "ID",
  ng = 2
)


keval <- function(df, model, age, ...) {
  sub <- paste0(
    "k = ",
    paste(model$call$ng, collapse = ""),
     "; Random = ",
    paste(model$call$random, collapse = ""),
    "; idiag = ",
    paste(model$call$idiag, collapse = ""),
     "; nwg = ",
     paste(model$call$nwg, collapse = "")
  )
  lcplot(
    df = df,
    model = model,
    age = age,
    sub = sub,
    ...
  )
  print(kableExtra::kable(ksum(model), caption = sub) \%>\%
    kableExtra::kable_styling("striped", full_width = F))
}

pred <- lcpred(prep_data, age_vars = c("Time"), fixcov = c("X1" = 0, "X2" = 0))

keval(prep_data, model1, "Time")
}
