% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/keval.R
\name{keval_apply}
\alias{keval_apply}
\title{Model Class Evaluation for Each Model in List}
\usage{
keval_apply(df, model_list, age, ...)
}
\arguments{
\item{df}{data frame used to produce model. lcpred output}

\item{model_list}{List of lcmem output to evaluate}

\item{age}{string of age variable}

\item{...}{arguments to base plots (eg. main = "Title")}
}
\value{
kable output providing class analysis and plot output
}
\description{
Provides n and frequency of each class and post prob of classes.
Also plots the model results using lcplot function
}
\note{
if you are using RMarkdown and want to output the kable output to html set
resutls = "asis"
}
